Rebase with the latest patches in a chain
Each commit gets its own history in Gerrit, so when fetching one change,
you may get obsolete versions of ancestors: they may have been merged,
abandoned or updated. Instead of trying to merge that, use the latest
version of only the active changes to reconstruct a chain of patches.

Remove the orginal vs applied patches distinction. The patches
"directory" now contains the equivalent of the previous "applied"
directory: the differences between the baseline and the compiled source.

Remove the "conflicts" page and files.

Fixes: #1